feat: разбивает код на модули
